Refactor api test scaffolding a bit
